Vendor assets to make offline development easier
We are already serving via GitHub pages which is backed by a CDN so
linking to remove assets isn't buying us too much (plus it affects our
page if they go down).

I leave the webfonts as a remote link since they are not required to
render the page in a mostly correct fashion (they'll fall back to system
fonts).

Also move the assets into /css and /js and remove the preprocessing of
sfbrigade.css as this isn't necessary if you use relative links.

Fixes #54